Represent FPS as a float (but still display it as an int).

+++ src/com/gregmcnew/android/pax/Emitter.java
@@ -60,7 +60,7 @@ public class Emitter {
 	}
 	
 	public void update(long dt) {
-		int fps = FramerateCounter.getFPS();
+		int fps = (int) FramerateCounter.getFPS();
 		
 		float elapsedLife = dt;
